requirements.txt
fastapi==0.115.8
gunicorn
wheel
uvicorn[standard]==0.34.0
pandas==2.2.3
python-jose==3.3.0
python-multipart==0.0.20
requests==2.32.3
urllib3==2.3.0
charset-normalizer==3.4.1
pycryptodome==3.21.0
pyotp==2.9.0
ipinfo==5.1.1
pystache==0.6.7
pyqrcode==1.2.1
pypng==0.20220715.0
psycopg2-binary==2.9.10
pdfkit==1.0.0
wkhtmltopdf==0.2
aiosmtplib==4.0.0
python-magic==0.4.27
eth-account==0.13.5
web3==7.8.0 
numpy==2.2.2
captcha==0.6.0
cryptography==44.0.1
asyncpg==0.30.0
python-dateutil==2.9.0.post0
orjson==3.10.15
aiofiles==24.1.0
httpx==0.28.1